kinetic-ingestor/requirements_gui.txt
# GUI Dependencies for The Kinetic Ingestor
# Install with: pip install -r requirements_gui.txt

# PyQt6 and extensions
PyQt6>=6.7.0
PyQt6-WebEngine>=6.7.0  # For rendering Markdown in HTML (optional, used in preview)

# Existing dependencies (should already be installed)
docling>=1.0.0
pyyaml>=6.0
rich>=13.0.0
prompt-toolkit>=3.0.0
ollama>=0.1.0
